+++ src/features/home/petition/answer/answer.props.ts
@@ -1,0 +1,7 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface AnswerFeatureType {
+  id: number;
+  setComment: Dispatch<SetStateAction<string>>;
+  comment: string;
+}
