+++ src/features/home/petition/comment/comment.props.ts
@@ -1,0 +1,7 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export interface CommentFeatureType {
+  id: number;
+  setComment: Dispatch<SetStateAction<string>>;
+  comment: string;
+}
